requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-crispy-forms==1.6.1
django-toolbelt==0.0.1
gunicorn==19.7.1
psycopg2==2.7.1
pytz==2017.2
requests==2.13.0
selenium==3.3.3
static3==0.7.0
whitenoise==3.3.0
